requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
